+++ src/main/java/com/acltabontabon/openwealth/services/custodyservices/customer/SingleCustomerReader.java
@@ -9,6 +9,7 @@ import com.acltabontabon.openwealth.models.custodyservices.Customer;
 import com.acltabontabon.openwealth.services.ReadCommand;
 import com.acltabontabon.openwealth.services.custodyservices.position.CustomerPositionStatementReader;
 import com.acltabontabon.openwealth.services.custodyservices.transaction.CustomerTransactionStatementReader;
+import com.acltabontabon.openwealth.types.DateType;
 import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.task.TaskExecutor;
@@ -24,7 +25,7 @@ public class SingleCustomerReader extends ReadCommand<Result<Customer>> {
     private final String correlationId;
     private final String customerId;
 
-    public CustomerPositionStatementReader positionStatement(LocalDate date, boolean eodIndicator, String dateType) {
+    public CustomerPositionStatementReader positionStatement(LocalDate date, boolean eodIndicator, DateType dateType) {
         return new CustomerPositionStatementReader(restClient, apiProperties, asyncExecutor, correlationId, customerId, date, eodIndicator, dateType);
     }